backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
APScheduler==3.11.0
asyncer==0.0.8
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
colorlog==6.9.0
cryptography==42.0.8
datasets==3.2.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
dspy==2.5.43
dspy-ai==2.5.43
email_validator==2.2.0
fastapi==0.104.1
fastapi-cli==0.0.7
fastapi-sso==0.10.0
filelock==3.16.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.9.0
greenlet==3.1.1
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json_repair==0.32.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.53.7
lxml==5.3.0
magicattr==0.1.6
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
numpy==1.26.2
oauthlib==3.2.2
openai==1.58.1
optuna==4.1.0
orjson==3.10.12
packaging==24.2
pandas==2.1.3
peewee==3.17.8
platformdirs==4.3.6
propcache==0.2.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
rq==2.0.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.23
starlette==0.37.2
ta-lib==0.6.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.0.0
webencodings==0.5.1
websockets==14.1
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.31
zipp==3.21.0
supabase==1.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9